Save models to half

+++ experiments/superhf/shf_iterative_v1/run_shf_iterative.py
@@ -220,7 +220,8 @@ def load_language_model(
         MockLanguageModel()
         if language_model_name == "mock"
         else AutoModelForCausalLM.from_pretrained(
-            language_model_name  # , torch_dtype=dtype
+            language_model_name,
+            torch_dtype=torch.bfloat16,  # For now, force half precision
         ).to(device)
     )
     if wandb.config.lora_r != 0 and wandb.config.lora_alpha != 0:
@@ -233,7 +234,8 @@ def load_language_model(
             task_type="CAUSAL_LM",
             fan_in_fan_out=False,
         )
-        language_model = get_peft_model(language_model, lora_config)
+        # For now, force half precision
+        language_model = get_peft_model(language_model, lora_config).half()
         language_model.print_trainable_parameters()
 
     print(f"Instantiated language model: {language_model_name}")
